move code into helpers
