+++ contracts/infinity-pool/src/testing/tests/sim_tests.rs
@@ -1,1 +1,3 @@
+mod helpers;
 mod sim_direct_swap_nfts_for_tokens;
+mod sim_swap_nfts_for_tokens;
